test: move web-worker test to core
