+++ eslint.config.js
@@ -100,7 +100,7 @@ export default antfu(
     files: [
       `docs/${GLOB_SRC}`,
       `packages/web-worker/${GLOB_SRC}`,
-      `test/web-worker/${GLOB_SRC}`,
+      `test/core/${GLOB_SRC}`,
     ],
     rules: {
       'no-restricted-globals': 'off',
