+++ test/core/src/web-worker/objectWorker.ts
