+++ test/core/src/web-worker/selfWorkerDep.ts
