--- test/web-worker/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "@vitest/test-web-worker",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "test": "vitest",
-    "coverage": "vitest run --coverage"
-  },
-  "devDependencies": {
-    "@vitest/web-worker": "workspace:*",
-    "vitest": "workspace:*"
-  }
-}
