--- test/web-worker/setup.ts
@@ -1,1 +1,0 @@
-import '@vitest/web-worker'
